Restructure into separate files for readability
