+++ Procfile
@@ -1,1 +1,1 @@
-web: gunicorn --bind 0.0.0.0:$PORT server:app -w 1
+web: gunicorn --bind 0.0.0.0:$PORT ncss_api:app -w 1
