+++ ncss_api/__init__.py
@@ -1,0 +1,14 @@
+from .app import app
+
+from . import (
+    core,
+    swagger,
+    ascii,
+    buses,
+    emoji,
+    convert,
+    astronomy,
+    syllables,
+    secret,
+)
+
