+++ ncss_api/app.py
@@ -1,0 +1,3 @@
+from flask import Flask
+
+app = Flask('ncss-apis')
